+++ RaspberryPie/config.py
@@ -1,0 +1,4 @@
+import configparser
+
+config = configparser.RawConfigParser()
+config.read("config.cfg") # edit this file for your specific configuration
